Modify for amiibrOS Project
Added a default configuration file amiibrOS_defconfig to be used
with `make amiibrOS_defconfig`. This default configuration
contains the bare-minimum of packages (will need to be updated) for
making an amiibrOS SD card.

Added a few of Adafruit's CircuitPython packages to the packages
list via utils/scanpypi. These are required for amiibrOS_defconfig.

Modified package/Config.in to show the newly added python packages
in the menuconfig menu.

Modified package/python3/python3.mk to allow for host openssl which
was needed to be able to add additional pypi packages during make.